postgresql modules: move connect_to_db and get_pg_version to module_utils/postgres.py (#55514)
* postgresql modules: move connect_to_db, get_pg_version to module_utils

* postgresql modules: move connect_to_db, get_pg_version to module_utils 2

* postgresql modules: move connect_to_db, get_pg_version to module_utils 3
